Use ES6 syntax for imports
