+++ example/index.html
@@ -11,7 +11,7 @@
   <script src="https://code.jquery.com/jquery-2.1.3.min.js"></script>
 
   <!-- This is what you need -->
-  <script src="../dist/sweet-alert.min.js"></script>
+  <script src="../dist/output.js"></script>
   <link rel="stylesheet" href="../dist/sweet-alert.css">
   <!--<link rel="stylesheet" href="../lib/themes/twitter/twitter.css">-->
   <!--.......................-->
