+++ lib/test.js
@@ -1,0 +1,3 @@
+var hopp = "hej";
+
+export default hopp;
