+++ package.json
@@ -26,6 +26,8 @@
   },
   "homepage": "http://tristanedwards.me/sweetalert",
   "devDependencies": {
+    "babelify": "^6.0.2",
+    "browserify": "^9.0.8",
     "glob": "^5.0.3",
     "gulp": "^3.8.11",
     "gulp-babel": "^5.1.0",
@@ -35,6 +37,9 @@
     "gulp-rename": "^1.2.0",
     "gulp-sass": "^1.3.2",
     "gulp-uglify": "^1.0.2",
-    "path": "^0.11.14"
+    "gulp-wrap": "^0.11.0",
+    "path": "^0.11.14",
+    "streamify": "^0.2.4",
+    "vinyl-source-stream": "^1.1.0"
   }
 }
